Move action icon control to service worker
